Fast pruning of modulestore MongoDB Structures
This creates the structures.py script which has "make_plan" and "prune"
commands (run the script with --help to get usage details). The overall
pruning logic is based on the work already done in prune_modulestore.py,
but this script makes significant changes to improve performance and
operational safety. It's been tested on a system with about 5 million
structure documents: 2 hours for make_plan, 8.5 hours to prune with a
batch size of 1000 and a delay of 5 seconds (--delay=5000).