Always get the source from the target github repo.
The problem we were running into is that there is a condition, where we were returing the private hash but not the public repo in the output yaml.

Looking at this code, we couldn't determine any situation where we would want to do that. The octopus_merge in this code ensures
that whatever hash we want is available in the target repo.