Correct test_get_pr_in_range mock for search_issues.
It was previously assuming that the only non-kwarg arguments wolud be
shas but we change how we do the search beacuse of a change in the
upstream api.  This means the search query has items in it that are not
shas but are needed to narrow the query in a real search.