Wait longer for ASGs to become healthy.
We launch a lot more machines than we used to so on slow ec2 days, this
can sometimes cause the deploy to timeout while waiting for all the
machines to exist and be healthy.  Instead of waiting 10 minutes, wait
15 minutes for all the machines to come in and become healthy.

This will mean it will take slightly longer to notice failures but it
should result in fewer failuers where we have no action but to just
re-deploy the app.