fix: switch send_email to use boto3
boto2 send_email() outright doesn't work anymore, and the package is
totally out of maintenance now so our only option is to migrate to
boto3.  This was brought to our attention when it started breaking the
user retirement pipeline (the hubspot stage which sends a notification
email internally).  The AWS error message was clear that we must upgrade
our API client (boto2).

I also ran make upgrade.