feat: script errors and exits if wrong version
Before merging approved prs into target branch, merge_approved_prs
script now looks to make sure the target branch's head is on correct
version.

This is to make sure nothing gets merged into target branch between the
start of this script and the point at which Prs are merged in.

This mainly affects edxapp pipelines and makes sure the version that
started the pipeline gets into production and not any subsequent commit.