feat: Add logging to diagnose race condition and branch behavior
I believe there's a race condition in the use of `--source-branch` here,
but I don't want to try fixing it until I understand the behavior of the
target branch -- this code should result in isolated commits hanging off of
the history of source-branch, but instead the commits form a merge chain.
I'll be able to compare the additional logging here against the actual
results in the git repos.

ref: ARCHBOM-1895