feat: Don't try to merge security patches when there aren't any
When there are no approved PRs, don't perform a merge -- just take the
source_branch commitish as the deployment SHA and new target_branch value.

When `octopus_merge` was called with an empty `approved_prs` list, it was
invoking the no-arg version of `git merge`, which merges the upstream
branch instead (as an implicit argument). So, instead of creating a single
commit off of each source-branch commit, the target-branch commits formed
a *chain* parallel to the source branch (each having two parents -- one a
source branch commit, and the other the previous target-branch commit.)

From `git help merge`

> If no commit is given from the command line, merge the remote-tracking
> branches that the current branch is configured to use as its upstream.

This was probably harmless but created a git history that was overly
difficult to understand.

Also:

- Rename `merge_sha` variable to `deploy_sha` to clarify intent, especially
  since it is now not always a merge commit
- Remove some logging that turned out to be redundant (logging had been
  added to figure out where that strange merge commit was coming from)

ref: ARCHBOM-1895