fix: reduced batch size due to userRateLimitExceeded
In the API Console, there is a similar quota referred to as Requests per 100 seconds per user. By default, it is set to 100 requests per 100 seconds per user and can be adjusted to a maximum value of 1,000. But the number of requests to the API is restricted to a maximum of 10 requests per second per user. You refer following document :https://developers.google.com/analytics/devguides/config/mgmt/v3/limits-quotas