feat!: Allow choosing deployment tag's name when merged private PRs
- Introduce new option `--target-tag` in `merge_approved_prs.py`
- Change the default base name of the tag from `release` to
  `release-candidate`
- Clarify existing help string

The existing script created tags that looked like `release-20230309135946`,
but these are made at the *beginning* of the deployment process and only
indicate an "intent to deploy"; not all of them will be released. They
also look very much like the `release-2023-03-03-10.42` tags, which
actually represent finished releases.

The new default will be clearer to operators. Allowing the pipeline to
set the full name will allow it to include the GoCD pipeline counter (or
similar), like the now-deleted `release-candidate-1234` branches in the
public repo, and will also allow the pipeline to choose its own date
format.

This is part of https://github.com/edx/edx-arch-experiments/issues/220